Put check_bigint func into db utils
