+++ src/amms.py
@@ -16,7 +16,7 @@ from solana.rpc.api import Client
 from solders.pubkey import Pubkey
 from solders.account_decoder import ParsedAccount
 
-from db import AmmLiquidity, get_db_session
+from db import AmmLiquidity, get_db_session, check_bigint
 from src.protocols.anchor_clients.bonkswap_client.accounts import Pool as BonkPool
 
 AUTHENTICATED_RPC_URL = os.environ.get("AUTHENTICATED_RPC_URL")
@@ -25,25 +25,6 @@ if AUTHENTICATED_RPC_URL is None:
 
 LOG = logging.getLogger(__name__)
 
-
-def check_bigint(value: int) -> int:
-    """
-    Checks if the given integer value fits within the PostgreSQL bigint range.
-
-    Args:
-            value (int): The integer value to check.
-
-    Returns:
-            int: Returns -2 if the value exceeds the PostgreSQL bigint range
-    """
-    # Define the bigint limits in PostgreSQL
-    bigint_min = -9223372036854775808
-    bigint_max = 9223372036854775807
-
-    # Check if the value is within the bigint range
-    if not bigint_min <= value <= bigint_max:
-        return -2  # Return -2 if the value is outside the bigint range
-    return value
 
 
 class Amms:
@@ -92,13 +73,6 @@ class Amms:
         fluxbeam_amm.timestamp = timestamp
         await fluxbeam_amm.get_pools()
         fluxbeam_amm.store_pools()
-
-
-# TODO: To be implemented.
-# def load_amm_data() -> Amms:
-#     swap_amms = Amms()
-#     swap_amms.update_pools()
-#     return swap_amms
 
 
 class Amm(ABC):
